pyproject.toml
[project]
name = "islr-calculator"
version = "0.1.0"
description = "Venezuelan Income Tax (ISLR) Calculator with friendly terminal UI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "questionary>=2.0.1",
    "rich>=13.7.0",
]
